+++ resources/views/components/docs/sidebar.blade.php
@@ -85,12 +85,7 @@
 
     <x-vellum::ui.scroll-area class="min-h-0 flex-1 px-4 pt-1 pb-4">
         <nav aria-label="Documentation" class="flex flex-col gap-1 text-sm">
-            @include('vellum::components.docs.partials.nav-tree', [
-                'nodes' => $navigation,
-                'activeSlug' => $activeSlug,
-                'containsActive' => $containsActive,
-                'depth' => 0,
-            ])
+            {!! \Vellum\Support\NavTree::render($navigation, $activeSlug, $containsActive) !!}
         </nav>
     </x-vellum::ui.scroll-area>